pyproject.toml
[project]
name = "rentalshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "ipykernel>=6.29.5",
    "langchain>=0.1.0",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.4.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "tabulate>=0.9.0",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "uvicorn>=0.24.0",
]
